adicionando calculadora de horas

+++ calculadoras_trabalhistas/calculadora_horasExtras.py
@@ -1,6 +1,6 @@
 print('**Calculadora de Horas Extras**')
 
-salario_bruto = int(input("Qual o seu salário bruto?"))
+salario_bruto = float(input("Qual o seu salário bruto?"))
 
 hora_normal = salario_bruto / 220
 hora_extra = hora_normal / 0.5
